algo-trader/requirements.txt
pandas>=2.0
numpy>=1.24
ccxt>=4.0
fastapi>=0.100
uvicorn>=0.23
pyyaml>=6.0
python-dotenv>=1.0
pytest>=7.4
requests>=2.31
jinja2>=3.1
python-telegram-bot>=20.0
websockets>=11.0
